requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
aiomysql==0.2.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.0.3